factor out command management
