+++ opendevin/agent.py
@@ -3,6 +3,7 @@ from typing import List, Dict, Type
 from dataclasses import dataclass
 from enum import Enum
 from .lib.event import Event
+from .controller import AgentController
 
 class Role(Enum):
     USER = "user"  # the user
@@ -63,7 +64,7 @@ class Agent(ABC):
         return self._history
 
     @abstractmethod
-    def step(self, previous_output: Event) -> Event:
+    def step(self, cmd_mgr: AgentController) -> Event:
         """
         Starts the execution of the assigned instruction. This method should
         be implemented by subclasses to define the specific execution logic.
