+++ opendevin/controller.py
@@ -9,13 +9,10 @@ class AgentController:
     def __init__(self, agent, max_iterations=100, callbacks=[]):
         self.agent = agent
         self.max_iterations = max_iterations
+        self.background_commands = []
         self.callbacks = callbacks
         self.callbacks.append(self.agent.add_event)
         self.callbacks.append(print_callback)
-
-    def get_background_logs(self):
-        # TODO: move and implement
-        return []
 
     def maybe_perform_action(self, event):
         if not (event and event.is_runnable()):
@@ -40,7 +37,7 @@ class AgentController:
                 for callback in self.callbacks:
                     callback(event)
 
-            action_event = self.agent.step()
+            action_event = self.agent.step(self)
             for callback in self.callbacks:
                 callback(action_event)
             if action_event.action == 'finish':
@@ -53,3 +50,42 @@ class AgentController:
                     callback(output_event)
             print("==============", flush=True)
 
+    def get_background_log(self, idx, cmd, stream, name):
+        logs = ""
+        while True:
+            readable, _, _ = select.select([stream], [], [], .1)
+            if not readable:
+                break
+            next = stream.readline()
+            if next == '':
+                break
+            logs += next
+        if logs == "": return
+
+        event = Event('output', {
+            'output': logs,
+            'stream':name,
+            'id': idx,
+            'command': cmd.args,
+        })
+        self.add_event(event)
+        return event
+
+    def get_background_logs(self):
+        all_events = []
+        for idx, cmd in enumerate(self.background_commands):
+            stdout_event = self.get_background_log(idx, cmd, cmd.stdout, 'stdout')
+            if stdout_event:
+                all_events.append(stdout_event)
+            stderr_event = self.get_background_log(idx, cmd, cmd.stderr, 'stderr')
+            if stderr_event:
+                all_events.append(stderr_event)
+
+            exit_code = cmd.poll()
+            if exit_code is not None:
+                event = Event('output', {'output': 'Background command %d exited with code %d' % (idx, exit_code)})
+                all_events.append(event)
+                self.add_event(event)
+
+        self.background_commands = [cmd for cmd in self.background_commands if cmd.poll() is None]
+        return all_events
