+++ opendevin/lib/actions/__init__.py
@@ -1,5 +1,5 @@
 from .run import run
-from .run import kill
+from .kill import kill
 from .browse import browse
 from .write import write
 from .read import read
